chore: move network api to utils
